requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
antlr4-python3-runtime==4.9.3
anyio==3.6.2
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
blessed==1.20.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cmake==3.25.0
contourpy==1.0.7
croniter==1.3.14
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.0
docstring-parser==0.15
fastapi==0.88.0
filelock==3.9.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
h11==0.14.0
hydra-core==1.3.2
idna==3.4
imageio==2.28.1
importlib-resources==5.12.0
inquirer==3.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonargparse==4.21.1
kiwisolver==1.4.4
lazy_loader==0.2
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
lit==15.0.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.2.1
multidict==6.0.4
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
opencv-contrib-python==4.7.0.72
opencv-python-headless==4.7.0.72
ordered-set==4.1.0
packaging==23.1
Pillow==9.3.0
protobuf==3.20.3
psutil==5.9.5
pydantic==1.10.7
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
readchar==4.0.5
requests==2.28.1
rich==13.3.5
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.9.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
sympy==1.11.1
tensorboardX==2.6
threadpoolctl==3.1.0
tifffile==2023.4.12

tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typeshed-client==2.3.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.5.1
websockets==11.0.3
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
